feat(utils): allow interactive file selection for git add
This commit adds functionality to the GitAdd function in utils/git.go, allowing for interactive file selection. This change requires modifications to the main function in order to comply with the new method signature. The commit also contains a modification to the main function in order to call GitAdd without the path parameter. Created by autommit 🦄.