refactor(utils): remove unnecessary code and fix argument in GitCommit method
The file utils/git.go was modified to remove the argument signCommits from the GitCommit method, since it was not used in the logic. Additionally, the file utils/utils.go was modified to remove a commented line. These changes were made to improve the code quality and maintainability.
Created by autommit 🦄.