feat(utils): add PgpKeyPath field to Autommit struct
feat(utils): add PgpKeyPath field to Autommit struct

This commit adds a new PgpKeyPath field to the Autommit struct in the utils/autommit.go file. This field will hold the path to the PGP key used to sign commits. It also updates the NewAutommit function to take in a pgpKeyPath argument and initialize the PgpKeyPath field in the Autommit struct.

Created by autommit 🦄.