feat: add conditional statement for PGP keyring in main.go
feat: add conditional statement for PGP keyring in main.go

Added an if-else statement for the case where no PGP key is provided. If the path to the keyring is not empty, the function to get the PGP keyring will be executed. Otherwise, a message will be printed to the console indicating that no PGP key was provided and commits will not be signed. 

Created by autommit 🦄.