docs: update usage instructions for Docker image
docs: update usage instructions for Docker image

Add a warning that the Docker image is currently unstable and recommend using the binary builds instead. This is to ensure that users are aware of the potential issues when using the Docker image. The README file has been updated accordingly. 

Created by autommit 🦄.