feat(cmd): add error check in git user info population
feat(cmd): Add error check in git user info population function in main.go file. 

Previously, there was no error check implemented in PopulateGitUserInfo function. This commit adds an error check which ensures that the program exits with a non-zero status code and prints the error to the console if there was an error with populating the git user info from .gitconfig. This improves the reliability and usability of the program. 

Created by autommit 🦄.