test(utils): add tests for FindDotGit and getDirectoryLevelsToRoot
test(utils): add tests for FindDotGit and getDirectoryLevelsToRoot

Add tests for the functions FindDotGit and getDirectoryLevelsToRoot. FindDotGit tests the function's ability to locate a .git directory in a given repository path. getDirectoryLevelsToRoot tests the function's ability to determine the number of directory levels to the root directory. Created by autommit 🦄.